requirements.txt
python-dotenv>=1.1.0
google-generativeai>=0.8.5